Refactor out the active extension icon
